requirements.txt
# VAEL – Python dependencies
# Install: pip install -r requirements.txt

# ── Core ─────────────────────────────────────
httpx>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
rapidfuzz>=3.6.0
pyyaml>=6.0.1
packaging>=24.0
cvss>=2.6                 # CVSS vector → base score (used by OSV CVSS parser)

# ── CLI ──────────────────────────────────────
typer>=0.12.0
rich>=13.7.0

# ── REST API ─────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
gunicorn>=22.0.0
aiofiles>=23.1.0          # required by StaticFiles for web UI serving
python-multipart>=0.0.9   # required by FastAPI file uploads

# ── AI Reasoner ──────────────────────────────
# Optional — pipeline falls back to deterministic verdict if not installed
google-generativeai>=0.7.0

# ── HTML parsing (Packet Storm, Seebug, intl search scrapers) ────────────────
beautifulsoup4>=4.12.0

# ── Testing ──────────────────────────────────
pytest>=8.0.0

# ── Future stages (4-6) ──────────────────────
# networkx>=3.3            # attack path graph
# stix2>=3.0               # ATT&CK STIX bundles
